Removed unused profile editor UI classes.
This was a work-in-progress which never really materialized, but the UI logic remained.
This logic is unreachable from Galicaster, and if made to be reachable by changing the code, it does not currently work.
I chose to remove it for code clarity. In case we had to work on this again,
  I believe having to search the git history or do a fresh start are preferred rather than keeping it on the code.