allows setting of pause dialog size in the conf.ini
our users had trouble with the tiny default pause button. they
often tried to click the disabled transport controls in the bottom
right to unpause and got confused when nothing happened.

to fix this we enlarged the pause button so that it is unmissable!

this adds that feature as a conf option:

```
[basic]
pausedialog_size = 70
```

the value is the percentage of the screen height you would like
the pause dialog to take up. 70 seems good for us.

this also scales the svg correctly to avoid a blurry icon.