improve plugin loading logging
logs stack trace when something goes wrong inside the plugin
(very useful when developing). also logs whether the plugin is
internal or external when it loads, which helps understand which
version of a plugin is loading (if you have installed plugin with
pip as well as in the galicaster plugins directory for example).