updates to checkrepo plugin. addressing issues #487 and #524 If galicaster restarts during a recording, on the next long heartbeat with an iCal update if the recording should still be happening it is resumed. once the recording is finished any track files belonging to the recording in rectemp are concatenated with ffmpeg and put into the correct mediapackage to then be ingested. there is also an option to delay the file concatenation and ingest until nightly time (this is to allow a recording in quick successtion to not wait for this job to complete)
[repository]
recoverytype = save

[plugins]
checkrepo = True

[checkrepo]
to_merge = True
delay_merge = False

its important to note some changes were made to galicaster behaviour around file recovery to accomadate this.
on galicaster restart autorecover 'full' recoverytype was always active, this has now been made configurable via the 'autorecover' option as with error handling. and for this plugin should be set to 'save'
only manual recordings should be given a prefix of "Recovered - " when autorecovered as this changes metadata for opencast scheduled recordings. maybe this information is better in the galicaster.xml if UI notification of a recovered MP is needed. this is not addressed here